amd64: mark io_in/out functions a public
It works to have the io_in/out functions in a private section of
the class, because the interface class Io_device declares those
functions as public and handle_io_access() only operates on Io_device
pointers to the actual derived Rtc class.

However, there is no need to have those function be private in the Rtc
class and doing so just creates questions.
Therefore, this commit changes the visibility of io_in/out() to public
and marks the warn/info/trace functions as private.

Change-Id: I73cc55b3a02cb55cbf90246a54e847e212829f84