add MSI support to PCI emulation
In difference to the existing MSI-X support, we need to emulate access
to the PCI configuration space, as MSIs are configured throught the MSI
capability of the PCI device.
Write accesses need to be emulated and when the driver enables MSI, the
uvmm configures the hardware device to match the guest's expectations.

Currently, we do not support multiple MSI messages, thus, we need to
emulate reads of the MSI.control register.

Thanks to Steffen Liebergeld and Jan Klötzke for their contributions to
this feature.

Change-Id: I1cb508c1c010679cca96add9193d721cd3c14bb0