arm: on vCPU entry, enable FPU and save the FPU registers
This change is required to make Uvmm work with code compiled with gcc-9
or later where floating-point instructions are part of the Uvmm binary.

When leaving the guest we have to
- enable CP12+CP13 for PL1 at CPACR to ensure that we have access to the
  FPU registers
- enable the FPU at FPEXC
- save the FPU/SIMD registers d0-d15 and maybe also d16-d31

We have to restore the entire FPU state (CPACR, FPEXC, all FPU/SIMD
registers) before re-entering the guest.

See ticket #CD-278.

Change-Id: I5f0f9b12c6e96e39d640bcc7cd6e73f818f5e218