PIC: correctly handle reads on even port
Depending on the last "Read Register Command" either the IRR or the ISR
register is visible on the even port. This command is issued by an OCW3
with RR=1. Once selected, the even port will keep on exposing the same
register. Reads from the odd port always return the IMR. To quote the
manual:

    There is no need to write an OCW3 before every status read
    operation, as long as the status read corresponds with the previous
    one; i.e., the 8259A ‘‘remembers’’ whether the IRR or ISR has been
    previously selected by the OCW3. This is not true when poll is
    used.

    After initialization the 8259A is set to IRR.

    For reading the IMR, no OCW3 is needed. The output data bus will
    contain the IMR whenever RD is active and A0 e 1 (OCW1).

Simplify the code a bit because there is no notion of a command any
more. Since we do not support poll command we also don't have to handle
the case when reading the even port.

Change-Id: I817afb7f4d4d17b2a181dd90fadd58180fc53bc8