pci: support PCI BAR map/unmap
The PCI BAR resources are registered only after the respective decoder
bits are enabled in the command register. If a decoder is disabled, its
resources are unmapped from the guest. While being disabled, the guest
can reprogram the BAR. To match the PCI spec more closely, all PCI
devices start disabled and no resources are initially mapped.

It is explicitly not supported to remap BARs that are currently
enabled. This is considered an error of the guest and actively
rejected. Such behaviour would also be undefined on real hardware and
all guests that do reprogram BARs (Linux, Tianocore) always disable the
device before.

Change-Id: Ia6cc39ec5138acacd3634f02ea44ac9f2cfb178a