Calculate order on every iteration
The order depends on the alignment of the start of the range and the
remaining size. This means the order can increase or decrease over the
given range. Therefore recalculate the order in every step. We have to
take both, the send and receive window into account.

This also fixes the problem where we would map behind the range end
because of the fixed order size.

Change-Id: Ife787b67b3b00e60d9d3f6cb710a9862d34cbd42