AArch32: Fix size assumption of floating point registers
We save the floating point registers when entering the uvmm. We
always save d0-d15 and save d16-d31 when necessary. We allocate
space for all registers on the stack with 32 * sizeof(d) and
assume sizeof(d) == 4.

The actual size of d-registers is 8 byte, so we do not reserve
enough space. This leads to access above the stack top which in
turn leads to an unresolvable page fault. The pagefault leads to
an exception which is not handled correctly by uvmm. It simply
resumes the previous state which in turn leads to a pagefault again.

Since we derive the entry stack pointer for the vcpu upcall from
the current stackpointer before the very first vcpu resume, the
problem is masked by printf leading to additional stack usage and
is only visible in release mode.

Change-Id: Ibe3417f7f7ed922dbde74cd1c8a292fc02cd52ab