amd64: add support for ACPI tables via Qemu_fw_if
If a firmware is run as initial guest, the ACPI tables cannot be
written directly into guest memory. Instead they are made available
through the Qemu_fw_if to be fetched by the firmware. To ease
integration, we provide the same interface as Qemu does.

The actual setup of the tables is a bit more involved because the final
location of the tables in memory is not known by the vmm in advance.
This is solved by the "etc/table-loader" command table that holds the
required relocation information for the firmware.

Change-Id: Ic2414990b6f4c1e6b546a0e3767669af5ff640d7