Handle invalid mmio accesses that hit a device correctly
If a handler for a mmio access is found, it could still be an invalid
access. In such a case the device handler will return an error code. We
want to apply the same logic to such kinds of failures as we already
do, when no hander is found at all.

Change-Id: Ia1ab9530830ebd9c83b7d3adf3db760fd95babb5