Rework PCI bridge device tree binding
Follow the device tree binding definition [1] more closely and remove
redundant code from the devices. Contrary to the previous
implementation the reg property of child nodes is interpreted as a
sequence of "addressable regions". In particular, each entry specifies
the BAR register and BAR type together with the BAR size. The setup of
the internal 'bars' array is now done in the generic Virt_pci_device
class based on this information.

The address of the reg entry is *not* the one stored in the BAR
register. This is represented by the "assigned-addresses" device tree
property. Right now this is not yet used and we fully rely on the
firmware or the kernel to assign proper BAR addresses in the bridge
window.

The bridge window is now solely represented by the ranges property. The
bridge setup code will parse it to amend the ACPI tables accordingly.
Also the handling of the first word of each ranges entry is fixed. The
spec states that only the type bits [25:24] are compared and the other
bits are masked.

[1] https://www.devicetree.org/open-firmware/bindings/pci/pci2_1.pdf

Change-Id: Icf9b8679d9f6e45fa4879e1f4e6470fa7b127d22