device_tree: add (E)ISA bus binding
The (E)ISA bus binding is defined in [1] and Linux actually supports
this binding. It can be used to express IOPORT resources in addition to
the regular MMIO resources. In a nutshell, it makes the first word of
the reg property a flags field, where bit 0 defines the IOPORT
property. To make this work, the device tree reg property getter is
amended with an optional Reg_flags output parameter that can receive the
meta data about the reg entry.

One thing that is special, is that in the "ranges" property only bit 0
of the first reg word is taken into account upon translating the
address. The other bits of the first word are masked.

[1] https://www.devicetree.org/open-firmware/bindings/isa/isa0_4d.ps

Change-Id: Ic767af7ab7e4c9c21704faa9ad6a4b3bb02594e3