uart_8250: register vcon irq on first write
Unfortunately there is no "enable" bit in any of the registers that a
guest may have to set. On the other hand, a guest may use the UART in
polling mode. If we would not register the vcon irq then the data ready
status bit would never be set. So use the first register write as
trigger to know that the UART is used by the guest.

Change-Id: I2c1bc6fae1288dc945a4c89be98eadfd9e2185fc