rtc: fix status registers C and D
Both registers are read-only. In status register C some interrupt flags
are located. We don't set them yet, but make sure they are cleared when
being read. Status register D only has one bit (VRT) that indicates the
validity of the RAM and time. It needs to be hard-wired to 1.

Change-Id: I2a7be87b631b49b1ccaed58433a573e83122d12d