Fix IA32_BIOS_SIGN_ID and IA32_PLATFORM_ID MSRs
On Intel CPUs the OS or firmware must read the IA32_PLATFORM_ID (0x17)
to match a microcode update. Because there won't ever be a
microcode update when running in uvmm, just return 0.

Related to this is the IA32_BIOS_SIGN_ID MSR which is used to query the
success of a microcode update and to get the currently installed
microcode update signature. The implementation was not actually
returning the information to the guest because there was a typo in the
register constant and because the Vmx_state handler intercepted the MSR
before Vdev::Microcode_revision had a chance to handle it.

Change-Id: I711f9677e20b2bd2a9d7fad522a572b2e7012cdb