Fix virtio_pci queue size handling
The PCI virtio binding handles the queue size differently than the MMIO
counterpart. In case of PCI there is only a single field (queue_size)
defined in the configuration structure. It is defined to specify the
maximum queue size on device reset. A driver *may* modify the value to
reduce the memory footprint but is not required to do so. Because
l4virtio is modeled like the MMIO virtio transport, we must initialize
the queue size of each queue to the maximum at device reset.

A previous change to virtio_console set the queue_size to the maximum
queue size on reset. Doing so is unspecified for the MMIO transport,
thus it's removed from reset().

Change-Id: I884c09793f67c70f5f3abcbc52889a8b6095b425