pci: rework BAR register handling
The guest will always work on a shadow copy of the BAR registers
maintained in the 'bars' array. By unifying the handling in Pci_device
of these registers, some special handling can be removed from the PCI
bridge code. It also fixes a bug where the 'bars' array and the real
BAR registers would get out of sync in case of 64-bit BARs.

Change-Id: I38b0ac6a890b2011ca918e31a6e6b0a64cd88041