arm: Fix default virt dts configs
Move the virtual devices to the beginning of the address space and the
memory behind it. Make it possible to auto assign the size of the
available memory to uvmm and the reg entry by specifying 0.

This fixes the problem of overlapping regions when the memory did get to
big.

Change-Id: I6979cb4c59303296bccb8568d43f63fa9e52883c