Flush device config
The virtio protocol spoken between uvmm and l4virtio services is l4
specific. Therefore cached memory for the config page is used. However,
uvmm is responsible to maintain cache coherency if the memory is
accessed by the guest uncached. Uvmm is doing this already in
virtio_dev.h, however the initial state of the config page wasn't
handled correctly.

Fix this by cache cleaning it before starting the guest.

Change-Id: I9b716d3ef1f3abe3668244495fa03f9804d36fa4