Do not use Ref_ptr with non heap memory
Ref_ptr calls the destructor when the ref count reaches zero. This does
not work with non heap memory objects like in the case of the
redist/sgir. Convert them to be allocated from the heap to fix all the
semantics.
This also removes the necessity from manually increasing the ref count.

Fixes ticket #CD-555.

Change-Id: I646ad01cbd583d0c8220bd48ae340074dbc32560