amd64: set activity state to active after Exec_halt
Even with pending interrupts, there is no guarantee that an event will
be injected with the next VM entry after the halt. Therefore, we cannot
be sure that the event injection will change the core's activity state
to active.

A VM entry with the 'Halt' activity state manifests itself with exit reason 33:
"VM-entry failure due to invalid guest state.", with the VM instruction
error field set to 5: "VMRESUME with non-launched VMCS".

Therefore we now set the state to Active unconditionally.

Change-Id: I97f1cd1f1bca7ce173af363e814e6200c99a81ea