amd64: handle VM-exit cases with valid IDT-vectoring info
According to the Intel SDM a VM-exit during event delivery can happen in
the following cases:
* Exception or NMI
* Task switch
* APIC access
* EPT violation
* EPT misconfiguration
* page-modification log full events
* SPP-related events  (Sub-Page permission)

All other VM-exits that are caused by event delivery are considered
processed and do not set the IDT-vectoring information field.

Furthermore, fiasco is using IDT-vectoring information in case an
interrupt is already pending at the vCPU when entering the kernel via
vcpu_resume_commit().

In all cases that uvmm handles already, we re-inject the event again
on the next VM-entry. All unhandled VM-exit reasons get a marker, but
stay unhandled.
For the Fiasco.OC emulation case, this is no change in behavior.

The obvious deficit of this solution is that the priority of the
IDT-vectoring event and the event causing the VM-exit is not compared,
e.g.exception takes priority over interrupt.
This implementation is also not consulting the APIC to check if a higher
priority interrupt pending in the meantime, which might necessitate
requeuing of the IDT-vectoring event.

Change-Id: Id72643410a9cef420b0cfec87ce66d0f3073dfc7