ACPI: DSDT PCI bridge description allocates less IO ports
The PCI bridge description in the DSDT allocated the whole port range
[0x0, 0xffff] to the bridge, except [0xcf8, 0xcff] which are the bridge
access ports.

This leaves no room for further dynamic allocations of IO ports. In
case of the pci=noacpi flag on the Linux command line, Linux parses the
APCI resources of the bridge and marks them as reserved. Because of the
noacpi flag, PCI resource allocation cannot happen through the ACPI
path and thus cannot use the properly assigned resource windows. In
consequence the VirtIO console device located at [0x800, 0x87f] cannot
allocate this range and is disabled by Linux.

This change reduces the bridge allocated IO resource to the bridge
access ports itself and a single bridge window between 0x8000 and
0xffff (inclusive).

The allocation issue does not arise, if pci=noacpi is not given. Other
ACPI devices seem to be able to allocate their IO ports in the ACPI
controlled IO port range in any case.

Change-Id: I03221627cf2022a86660dab4b893151c49f1fe9e