amd64: update list of VMX exit reasons and stringifications
The list of exit reasons was incomplete and the maximum number of exit
reasons changed.
This updates the constant definitions to match the Intel SDM dated
12/2021, Vol. 3 Appendix C.

Furthermore, I fix a potential out-of-bounds access to the
str_exit_reason array, provide a safeguard for the future in form of a
static_assert and add all missing stringifications.

Change-Id: I6b5a41823486243bf9895858a3072b2d058a473a