device: Add virtio-input based on L4Re::Events
L4Re provides the L4Re::Event interface that is very similar to the
Linux input event API. At the moment there is no way to pass these
events into a VM, e.g. when the VM is shown inside a Mag window.
Fortunately the virtio-input device is also based on the Linux input
event API, so translating from L4Re::Event to virtio-input is straight-
forward for the most part.

There are two differences that need to be handled:

  1. virtio-input always refers to a single input device, while
     L4Re::Event might bundle multiple "streams" (e.g. mouse and
     keyboard). This is handled using an "event demultiplexer" that
     passes arriving input events to one of multiple virtio-input
     devices in uvmm.

  2. L4Re::Event allows hotplugging input devices, while virtio-mmio
     does not support that at the moment. All device information must be
     available before the VM is started. The implementation currently
     waits on startup until all the input devices show up. Depending on
     the L4Re::Event server this might be immediately (usually) or only
     once some input events have been handled (mag, at the moment).

The implementation is generic and works for all kinds of input devices:
mouse, keyboard, touchscreen, ...

Change-Id: I0b6d1e99c7aa45c80271d08b9083b314fee1b383