amd64: enable KVM clock feature bit for stable clocksource
The set feature bit does not promise that the KVM clocks are stable.
The set feature bit defines Vcpu_time_info.flags[0] as TSC_STABLE bit.
If set, the TSC_STABLE bit promises that all TSCs of all KVM clocks
across all cores are in sync.

At the moment this is not the case, thus Vcpu_time_info.flag[0] is
not set.

Change-Id: Icfdf4259c3155b0cb7497b2f7959db1588343d49