Support read-only MMIO regions
Map MMIO regions that have the read-only flag set with RO rights into
the guest.

In addition, read-write MMIO regions are now mapped with RW rights
instead of RWX rights. This is a more sensible default.
There might be use cases for executable MMIO regions, however that is
expected to be a rare case, thus it should be handled by adding an extra
device which explicitly maps with RWX rights.

Co-authored-by: Adam Lackorzynski <adam@l4re.org>
Change-Id: Ieafcbe36159fdf040454b5b97ffa6987531122a8
Requires-Change-Id: I41498866450c7a56a9cdd737ce96f397a9327b2c
Requires-Change-Id: If35a58780389bd9bcc7c226fb2aa5e9dd5e63a74
Requires-Change-Id: Iafd5d573900c352ee8676a7b1c34967bcc28ee27