Make binary loading more modular
Move the code to load images out of the guest class to self contained
modules. This makes it easier to maintain them and also allows to
enable/disables them depending on the use case.

The raw loader gained an optional parameter for the address the raw
image should be loaded to with the syntax: "raw:addr=0x????????:filename".

Change-Id: I2fb9f48ba3ce2a1d5811c519be15517b715bb5bd