x86/kvm_clock: do not update periodically
KVM_CLOCK (or PV_CLOCK) works by using the CPUs TSC as timer
source. To that end the shared data structure contains
everything the guest needs to get the current time; An initial
TSC value; the timestamp of the host when the TSC was
valid and the scalers. Updating the TSC is not required
at runtime except for situations where the host TSC changes
(e.g. during suspend/resume).

Change-Id: I72f356e9fa3d859968591806a31ae49b4ff9c959