x86/x2apic: support SELF_IPI msr
Some operating systems use self-IPIs --Inter-processor
Interrupts sent to the CPU that triggered the Interrupt--
extensively. For the X2Apic Intel introduced a special MSR to
trigger self-IPIs. The SELF_IPI MSR is a write-only MSR and
any read will result in a #GP.

Handling and prioritization of a self-IPI sent via the
SELF_IPI register is architecturally identical to that for an
IPI sent from the ICR.

We have seen Linux using this register and therefore we must
implement it instead of ignoring it.

Change-Id: I5732a8c582f3b94f9d62d1e04e84340eee65de3d