Add firmware boot support
Add a loader for ROM code. The loader starts the guest code at the
address provided. The --kernel parameter format is "rom:addr=0xfffffff0:".
The actual code has to be loaded by the ROM device.

In ROM mode no zeropage is set up and the code is started in real mode.
This first boot stage needs to setup the system in ways for the actual
second boot stage to successfully load the guest OS. The required info
can be provided e.g. via the qemu fw interface.

Change-Id: I8040e08a146f87d351906c2ef8477ce51d43b8ad