x86: disable hardware coordination feedback capability
ACPI defines a number of P-States that the OS can use to
control power utilization of the CPU. When multiple CPUs are
in the same package, there may be dependencies that the OS
must consider. In order to provide the OS with the required
information, Intel provides the P-State hardware coordination
capability, which is accessed with the MPERF/APERF MSRs. See
the intel programmer manual chapter 14.2 for reference.

We currently do not support P-States. Thus, this change filters
the P-State hardware coordination capability from cpuid.

In practice Linux showed a number of warnings when it tried to
access these MSRs and received #GPs. With this change Linux
does not see the capability and thus does not try to use these
MSRs.

Change-Id: Ie6cffbb59ce990d74b95bbbf30fc0b05b52b80e5