x86: enable PAT and MTRR
We preload the guest PAT with the reset value and make sure the value is
loaded on VM entry. The MTRR is hard-wired to zero.

This change is required for hardware pass-through.

The PAT was introduced with the Pentium III and is enabled
when the CPU supports it. CPUs older than the Pentium
III are unsupported because they do not sport vmx. Therefore
we do not support the absense of PAT.

Change-Id: Ide46ee2927f32aa83392aee4116a2b3665f862e0