Workaround fast string bug in the Linux kernel
As of 5.17, the Linux kernel operates under the following assumptions:

* The FSRM bit being set implies that the ERMS bit is also set in the
  Extended Features flags of CPUID.

* If the Fast String Enable bit in IA32_MISC_ENABLE MSR (0x1a0) is not
  set, the Linux kernel clears the ERMS bit.

The failure to also clear the FSRM bit when Fast String Enable is not
set by Linux causes a page fault in the implementation of __memmove on
CPUs that actually have the FSRM bit set (e.g. 11th Gen Intel Core i7).

While the logical error should be fixed in the upstream Linux, we
workaround the issue for existing Linux kernel releases by enabling the
Fast String Enable bit in IA32_MISC_ENABLE MSR. It is an architectural
bit and therefore it should be safe to enable it unconditionally.

Change-Id: I9a4fb896153bbb68892c94516f8e252341ae7191