cfi: add Intel Primary Algorithm Extended Query Table
This is needed to use the CFI flash emulation in Linux. For CFI flashes
that report the Intel command set it expects that the "Intel Primary
Algorithm Extended Query Table" is also present, otherwise it fails
with:

[    1.785858] 4000000.flash: Found 1 x8 devices at 0x0 in 8-bit bank.
Manufacturer ID 0x000000 Chip ID 0x000000
[    1.802248] gen_probe: No supported Vendor Command Set found
[    1.811493] physmap-flash 4000000.flash: map_probe failed

There is no need to fill in anything except for the header, since most
of the fields just specify support for additional features that are not
supported (and likely not useful) for the CFI emulation.

Change-Id: I22230bdafa645f231e6d4f3a5755c2826b8d03a3