cfi: add virtio-block backend
Optionally persist the flash content directly into a virtio-block
device. This is another option for persisting the flash memory. The
other option being a dataspace.

The virtio-block backend allows us to directly use a partition on an
ahci- or nvme-disk as persistent storage (using the ahci-drv and
nvme-drv respectively). With the dataspace backend, persistence required
an additional component.

Change-Id: Icb12d83545a173aba02773c48e81a74e59847dbb