cfi: only clear bits on writes like a real flash
At the moment the CFI emulation copies the bytes/words written directly
to the dataspace. While this works for most use cases it does not match
the behavior of a real (physical) flash. Typically, writes can only
clear bits. To set bits it is necessary to use the erase commands.

The CFI driver in Linux makes use of this property when writing a single
byte to flashes with larger bus widths: To write 0x42 to a 32-bit flash
it would write 0xffffff42 to the flash (and expect that only the 0x42
takes effect).

Make the CFI emulation work similarly using a bitwise AND operation
for flash writes. Note that this requires the flash driver to erase
the flash before initial usage (just like on a real flash).

Change-Id: I8bc741eb7be7e660b906911bb0cad8b329e34a1a