ARM: Fix SCTLR reset value for AArch32
At the moment only 3 bits are written to the guest SCTLR when a CPU is
reset on ARM: C, I (enable data/instruction cache) and CP15BEN (enable
CP15 barriers). This should work fine for AArch64 guests because the
reset value of the register is defined to be "architecturally UNKNOWN"
and the guest is expected to fully initialize the register.

However, in AArch32 state some bits have a defined reset value.
Guests might rely on this by updating only some bits using
read-modify-write operations. At least EDK2/Tianocore/UEFI is
affected when running in AArch32 state on QEMU: It fails shortly
after enabling the MMU if the SCTLR is not initialized correctly.
In AArch64 state it works without problems.

Change the initialization to set the bits defined in the ARMv7 reference
manual, but leave it as-is for 64-bit guests. However, for consistency
we can drop CP15BEN for 64-bit guests: It would only affect 32-bit
applications running in EL0, so it is better to leave it up to the guest
OS to initialize it as needed.

Change-Id: I3cf2ed726c2ed8bfd7cb32fa0667eb8e0517d956