x86: implement support for NMIs
Contemporary Linux kernels use the NMI to self-test IPI functionality.
This is done even on single core boots. Previously the NMI case was not
handled which resulted in the wrongly injection of an interrupt with
vector 0 which in turn is interpreted by the guest as #DE fault.

While we are at it, we also explicitly drop SMI requests. It's not
implemented at all and would also be injected incorrectly like the NMI
before.

Change-Id: I8f86509c5d88072a4bca3a8815012b9b66650ded