virtio_console: add error message to chksys
Some Vcon implementations, e.g. Moe's, do not implement the Icu
interface and if you pass such a cap to UVMM, creation of a VirtIO
console will fail at `attach_con_irq()` and throw an exception with no
error message.

Since it is not immediately obvious what went wrong, this commit adds
an error message to the chksys function that generates said exception.

Change-Id: If2005ec2c338a8b9e6ca31bc4287978188904c15