don't try to boot an ARM EFI module but complain
If the image header is an EFI header, don't try to boot it but show a
helpful message. This happens with newer binutils stripping the ARM
header and this can be prevented by adding 'nostrip' to the modules.list
entry.

Also don't boot images with 'MZ' header, maybe it was not possible to
interpret the EFI signature.

See ticket #CD-723.

Change-Id: I46cec4f7be27f133b7465529df46dd4ed684a5bd