amd64: Unify generation of ACPI tables
The previous approach to generating ACPI tables was rather limited in
its flexibility, all sizes and table references had to be pre-computed,
which gets particularly tricky with dynamically-sized tables.

In addition there was quite a bit of code duplication between Bios_ables
and Acpi_tables (Qemu_fw_cfg), which made adding or modifying ACPI
tables complicated and error-prone.

Therefore, this change refactors the generation of ACPI tables, so table
reference fields and checksum fields are not filled in immediately, but
instead a list of fixups is kept for them. This allows having the logic
for generating the ACPI tables in one place (Tables), instead of three
places (Tables, Bios_tables, Acpi_tables).

Change-Id: I2a90df6dd9cbfc90d562b9edfa2c2f8b0620f25c