device: framebuffer: Add format to device tree
On device tree platforms like ARM the framebuffer format (width, height,
stride and format) are read directly from the "simple-framebuffer" node.
Update the uvmm device tree with the values obtained from the Goos
framebuffer server to allow the guest to use the framebuffer correctly.

Only a small set of formats is currently defined in the (Linux) device
tree binding documentation for "simple-framebuffer" [1]. A framebuffer
with different format might still be usable e.g. on amd64 through the
"zeropage", so just warn if an unsupported format is detected.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/devicetree/bindings/display/simple-framebuffer.yaml

Change-Id: I37467a4edbfc3c1e748696ed09f251d6870457d5