amd64: framebuffer: Map only available framebuffer size
There are two different sizes in the framebuffer device:
  1. The size of the framebuffer dataspace.
  2. The "reg" size of the "simple-framebuffer" device tree node.

At the moment these two sizes are used inconsistently: Only the actual
size of the dataspace is attached to uvmm's address space, while the
full "reg" size is mapped into the guest. In some configurations this
can lead to warnings from Fiasco or inadvertent mapping of unrelated
VMM memory regions into the guest.

Fix this by ignoring the "reg" size completely and only make use of the
actual dataspace size for all mapping operations.

See ticket #CD-842 for a detailed example.

Change-Id: Idb71778e239b24b19363dc83c0e12adbfac6cbfc