device: Make framebuffer independent of ARCH-amd64
Most of the simple-framebuffer device implementation is also useful for
other architectures. Only the "Zeropage" related code is specific to
AMD64, since device tree platforms like ARM can simply detect the
framebuffer from the device tree provided by uvmm.

Make it possible to share most of the code by moving most of the device
implementation out of ARCH-amd64. The "Zeropage" code is kept in
ARCH-amd64 and called through a function in the Guest header.

Change-Id: I6d82a03b69160f47a6ef3c17f918d717e2a452bd