remove dma_space.associate() calls
The code wants to acquire the host-physical address to map the
guest RAM to the same address. This is unnecessary, because the
uvmm can set up the guest RAM in an DMA space to allow it to
perform DMA accesses, if there are DMA capable devices.

The Mmio_proxy also wants to know the host-physical address
to program the device accordingly. Currently it is unclear to
me, why we need to do this. If the feature is used, I throw an
error.

Requires-Change-Id: Ia646e590432b63441860f2249d4f076ca0bf7608
Change-Id: If0b6ebc56178d17c6fd5cd841f5a151bdd30cb97
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>