[API] l4re/dma_space: separate privileged part of interface
Move the L4Re::Dma_space::associate interface into L4Re::Dma_space_mgr.
This new interface is provided as a singleton by Moe and is now used by
Io as a privileged interface to manage the Dma_space associations for
the clients.

This removes complications with potentially dangerous use of the
Dma_space interface, that each Moe client can use as it likes.

Change-Id: Iea10ac3ff22fea54b7a14ea35af7e98fcadc9a1f
Signed-off-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>