fix off-by-one bug in 64-bit PCI BAR parsing
In case of a 64-bit BAR the code takes preventive measures to read
past the last BAR depending on the PCI config space type.
The code compares the offset of the last BAR with the offset of the
next BAR to be read. The comparison for greater-or-equal is flawed,
as the last BAR is valid.

Use greater-than comparison instead.

Change-Id: Iee40a1608abbac0826b5123b9b543a7713f49b81