amd64: framebuffer: Round up frame buffer size
The framebuffer size might not be a multiple of 64 KiB chunks. In this
case the validation of the framebuffer region might fail in the guest
because the size truncated to 64 KiB chunks is not enough to hold the
entire framebuffer.

Round up the size to ensure it is large enough to hold the entire
framebuffer.

Change-Id: Ie7afb1f514a24f8215c302c1bb8524124f8a648e