amd64: Reimplement instruction decoder
The previous iteration of the instruction decoder had many issues:
- TODOs, some of them documented some not, for example the instruction
  decoder did not correctly decode a disp32 specified in the SIB.
- Code duplication, generous use of magic constants and absence of
  comments made the code incomprehensible and thus difficult to
  maintain.
- Assumptions made by the decoder not stated.

Therefore the decoding was almost completely reimplemented. Decoding is
now subdivided into the different components an instruction can consist
of.

Furthermore, the decoder distinguishes between the address of the buffer
from which it reads the instruction and the instruction pointer of the
guest. This is necessary because all addresses determined by the
decoder for the operand are in guest virtual address space.

Change-Id: I3ca67d7420c912c93e27529b5554bcbd3d5bbd0c