Implement IOAPIC
Implement basic reading/writing IOAPIC registers.
This allows the guest to set up redirection tables and
makes the IOAPIC a valid choice for an interrupt handler.

This also changes the interrupt handler for uart to
IOAPIC. The PIC does not work with OpenBSD here, as
OpenBSD defaults to IOAPIC as soon as it is available.

Tested with Linux and OpenBSD.

Change-Id: If6b9161c90140fbe78b78028a4f438b2152aaf3b