framebuffer: initialize view_info to prevent warning
The compiler may warn that view_info is not initialized. This is a false
positive warning because the compiler cannot know what happens on the
server side.

The compiler does not warn if VMM::Generic_guest::register_framebuffer()
is overridden because in that case it assumes that the overridden method
initializes the 'info'. Therefore the compiler warned only on certain
architectures.

As this code is not performance-critical, initialize 'fb_viewinfo' to
prevent the several related warnings.

Change-Id: I71a09cd665aedb481f5013be1e4fb9e5a1925bd9