amd64: Specify revision for each ACPI table
So far, with the exception of the DSDT table, we just set the revisions
of all ACPI tables to zero. But the revision of a table can affect the
fields present in the table or other even other parts of ACPI, and thus
the parsing on the client side.

Therefore, it is now mandatory to explicitly specify the desired
revision for each table.

The revisions of the individual tables were set according to the ones
provided in section 5.2 of the ACPI specification version 6.3.
For DSDT, revision 1 was chosen instead of 2 in order not to change the
integer width of the AML code/interpreter. Whether this could be set
to 2, i.e. 64-bit integer width, without breaking the hardcoded AML
structures in Pci_host_generic would have to be determined first.

Change-Id: Ic0f4bd51312973bdb877bf160db218473da52912