pci_host_ecam: fix interrupt-map-mask mask entry
The first item of the 'interrupt-map-mask' entry contains a mask which
is left-shifted by 11 and the applied to the device ID of the PCI device
which is about to be attached to the ECAM bridge.

The dev_id=0 is used for the ECAM bridge itself, so dev_id=1 is the ID
for the first PCI device. This results in a value of 0x800 which is then
logically ANDed with the first item of 'interrupt-map-mask'. Hence, bit
11 should be also set there -- otherwise setup_device_irq() complains
with ``PCI device not found in interrupt map.''

Fix the comment to give a working example for the developer.

Change-Id: If85c2f2e563102ed577a94cb3c86a6e2bfb8951f