pci_virtio_device: provide IO and MMIO BAR configuration option
We support two configurations: Either the DT node contains a MMIO32 and
IO BAR each, or two MMIO32 BARs. The first MMIO32 BAR is always used for
the MSI-X table and the PBA, the second BAR contains the VirtIO
configuration space.

When the DT node features an MMIO32 BAR for the virtio configuration,
the BAR shall have a size of 4KB to adhere to PCI standards. However,
the implementation does not use the full 4KB, but limits the VirtIO
device specific configuration space to 512 bytes.

Change-Id: I1130ab6cf1dec7e0312f1b48805c91d6cac075de