Pci_conn: fix faulty assignment to out parameter in io_in().
The assignment of the value read from the virtio config space to the
out parameter did an additional bitwise-and. This is harmful, if the
out parameter is initialized with a value different than -1UL.

The only user of this function is the exit handling for IO port
access in the amd64 architecture code. This code path initializes
the out parameter with -1UL and thus the bitwise-and + assignment
is not harmful.

The change also now calls the Mem_access function to control the width
of the read value.

Change-Id: Ia425377e198892c08419d73ed5b884bd3fc79d3e