virtio-pci: extend PCI connector with access via MMIO BARs
At the moment, we only use IO port access for the VirtIO-PCI transport.
The same PCI BAR layout can be used with MMIO access in a MMIO BAR.
This commit generalizes the base class and adds an MMIO interface
next to the IO port interface.

The VirtIO devices can be configured from the device tree to
use MMIO32 or IO port based configuration access. This depends on the
presence of either an IO BAR or a second MMIO32 BAR in the reg
property of the DT node.

Configuring both, a second MMIO32 BAR and an IO BAR is not supported.

Change-Id: If9fe5e72162c69569c4c68b7c13d5a9d6d77c0bb