Pci_virtio_device: make hardcoded BAR expectations more dynamic
Up to now, we expected the MMIO BAR for the MSI-X table to be BAR 0 and
the IO BAR for the VirtIO-PCI transport to be BAR 1. This was hardcoded
in Virtio_device_pci.
This commit uses the requirements to the MMIO and IO BARs to detect the
first fitting BAR at run-time dropping the need for a strict register
order in the device's DT node.

Change-Id: Id5237a5f807060ea0c97688d01e2dea0eef5c6bf