connect IOAPIC to MSI distributor
The IOAPIC translates the Redirection Table Entry to an MSI addr/data
pair and sends it to the MSI distributor.
Unfortunately, the formats are not binary compatible, so we have to do
some translation effort. We gain that we do not have to implement a
second interrupt path including the addressing schemes.

This commit also takes destination mode and delivery mode into account
when delivering an interrupt.

Change-Id: Ic3d46671da6b283e8d25afc78bdb7b887a8dbef3