amd64: Avoid unnecessary interrupt window intercepts
Prior to this change, whenever the guest causes a VM exit while it has
interrupts disabled, it is quite likely that another unnecessary VM exit
is triggered shortly after the guest resumes.

What happens is that if the guest has interrupts disabled, the
can_inject_interrupt() method returns false, which leads uvmm to
unconditionally enable the interrupt window intercept. As soon as the
guest reenables interrupts, the interrupt window intercept triggers,
causing a VM exit. uvmm finds that there nothing to do (no IRQ pending
in the APIC), and then justs resumes the VM.

Such unnecessary VM exits have been observed, for example, when the
guest performs some IO port or MMIO accesses to the LAPIC
(interrupts are disabled during that).

To avoid these unnecessary VM exits, this change adds an
is_irq_pending() check, so that the interrupt window intercept is only
enabled if the guest has pending interrupts.

Change-Id: I39f62f874d81a8c57ce6c2d827371b58c65f06b9