amd64/pt_walker: rework debug output
The PT walker is only relevant in context of an access to an emulated
MMIO device. Change the Dbg component accordingly and set a subsystem
string.

Furthermore, add a compile time switch to most trace lines to reduce the
amount of output in trace levels. This allows to use Dbg::Trace again
instead of Db::Extensive.

Lastly, print the exception message thrown in the walk() method when
it's caught by the caller.

Change-Id: Id73c296d1837918d4e5f4b5fe6f98bbbcadf652c